+++ ironelder/AndroidArchitecture/app/src/main/java/com/ironelder/androidarchitecture/data/repository/SearchDataRepoImpl.kt
@@ -4,7 +4,7 @@ import com.ironelder.androidarchitecture.data.RetrofitForNaver
 import com.ironelder.androidarchitecture.data.TotalModel
 import retrofit2.Call
 
-object SearchDataRepoImpl:ISearchDataRepo<TotalModel> {
+object SearchDataRepoImpl : SearchDataRepository<TotalModel> {
     override fun getDataForSearch(type: String, query: String): Call<TotalModel> {
         return RetrofitForNaver.searchApi.requestSearchForNaver(type, query)
     }
