--- ironelder/AndroidArchitecture/app/src/main/java/com/ironelder/androidarchitecture/view/ICustomItemView.kt
@@ -1,5 +1,0 @@
-package com.ironelder.androidarchitecture.view
-
-interface ICustomItemView<T>{
-    fun setData(item:T)
-}
